[hotfix][table-planner] Move EnrichedRowData to table-runtime
